requirements.txt
tksvg
aiohttp
python-dotenv
requests
aioconsole
python-chess
ttkbootstrap